cpp/2019/04: rename files and Bazel targets
