+++ cpp/year2019/day04/internal.h
